% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_package}
\alias{check_package}
\title{Check if a package is installed}
\usage{
check_package(package)
}
\arguments{
\item{package}{The name of the package to check.}
}
\value{
A logical value: TRUE if the package is installed, FALSE otherwise.
}
\description{
This function checks whether a package is installed. If the package is not installed, it informs the user that the package is missing.
}
\details{
This function checks whether a package is installed. If the package is not installed, it informs the user that the package is missing.
}
\examples{
check_package('dplyr')
}
\author{
Sora Yoon, PhD
}
