% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_hic_file.R
\name{get_example_hic_file}
\alias{get_example_hic_file}
\title{Example download}
\usage{
get_example_hic_file(destfile = "example.hic")
}
\arguments{
\item{destfile}{Path to save the example Hi-C file}
}
\value{
Hi-C file download at destfile.
}
\description{
This function downloads an example Hi-C file from chromosome 19 of mice (mm10) Th1 cells.
}
