% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getElbowPoint}
\alias{getElbowPoint}
\title{Estimation of elbow point from J-shaped-curve}
\usage{
getElbowPoint(numbers)
}
\arguments{
\item{numbers}{Numeric vector}
}
\value{
A \code{list} containing two elements: \code{index}, the index of a point in a sorted vector of numbers in descending order, representing the point where the tangent is closest to one, and \code{ConnectivityCutoff}, the corresponding value at that point.
}
\description{
It provides a point of the highest curvature from a J-shaped-plot
}
\details{
Estimation of elbow point from J-shaped curve
}
\examples{
modulefile = system.file('extdata','mouse_naiveCD4T_Vahedi_short.rds',
package = 'HiCocietyExample')
mycom = readRDS(modulefile)
connec = mycom$ModuleSummary$connectivity
getElbowPoint(connec)
}
\author{
Sora Yoon, PhD
}
