% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hic2network}
\alias{hic2network}
\title{HiC to network data format}
\usage{
hic2network(ftab)
}
\arguments{
\item{ftab}{three-column data composed of locus1, locus2 and value}
}
<<<<<<< HEAD
\value{
An \code{igraph} object representing statistically significant chromatin interactions.
}
=======
>>>>>>> 3ed519ee94c9f81647aee196f9f42b73ceff4987
\description{
It converts Hi-C dataframe to network object.
}
\details{
Convert HiC to network data format
}
\examples{
<<<<<<< HEAD
# This example might take a long time to run, so we wrap it in donttest{}
\donttest{
myhic=system.file('extdata', 'example.hic', package ='HiCocietyExample')
ftab=getContactFrequency(myhic,19,5000);
net = hic2network(ftab[1:100,]);
plot(net)
}
=======
# Not run
# myhic=get_example_hic_file();
# ftab=getContactFrequency(paste0('./',myhic),19,5000);
# net = hic2network(ftab[100:200,]);
# plot(net)
>>>>>>> 3ed519ee94c9f81647aee196f9f42b73ceff4987
}
\author{
Sora Yoon, PhD
}
