% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{install_and_load}
\alias{install_and_load}
\title{Pakcage install and load}
\usage{
install_and_load(package)
}
\arguments{
\item{package}{package name}
}
\description{
It loads a package.
}
\details{
Loading a package
}
\author{
Sora Yoon, PhD
}
