% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{maptocolors}
\alias{maptocolors}
\title{Numbers to color vector}
\usage{
maptocolors(vec, col, num = 100, range = NULL)
}
\arguments{
\item{vec}{Numeric vector to be converted to the color vector}

\item{col}{Colors to be mapped.}

\item{num}{Length of sequence where the min and max are those from the truncated / or the original vectors}

\item{range}{Truncated values' range}
}
\description{
It maps numbers to colors.
}
\details{
Numbers to color vector
}
\author{
Sora Yoon, PhD
}
