% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RankDiff}
\alias{RankDiff}
\title{Rank difference between two conditions}
\usage{
RankDiff(wt, ko)
}
\arguments{
\item{wt}{hic2community result from condition 1}

\item{ko}{hic2community result from condition 2}
}
\description{
output table of rank difference of modules between cell types is generated.
}
\details{
Rank difference between two conditions
}
\examples{
# Not run
# RankDiff(mycom1, mycom2)
}
\author{
Sora Yoon, PhD
}
