% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calculate_avg_count}
\alias{calculate_avg_count}
\title{Calculate Average Count within 5-pixel Padding}
\usage{
calculate_avg_count(x, y, counts, resol)
}
\arguments{
\item{x}{Numeric vector of x-coordinates of contact frequency data frame.}

\item{y}{Numeric vector of y-coordinates of contact frequency data frame.}

\item{counts}{Numeric vector of contact frequency counts.}

\item{resol}{Integer specifying the HiC resolution.}
}
\value{
A numeric vector of average counts.
}
\description{
This function calculates the average count within a 25kb padding
around each (x, y) coordinate pair.
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(1, 2, 3, 4, 5)
counts <- c(10, 20, 30, 40, 50)
resol <- 10000
calculate_avg_count(x, y, counts, resol)
}
