% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_all_chr_names}
\alias{get_all_chr_names}
\title{Retrieve chromosome names from .hic file}
\usage{
get_all_chr_names(fname)
}
\arguments{
\item{fname}{Path to .hic file}
}
\value{
A character vector containing the names of chromosomes whose genomic lengths exceed 2.5 Mbp.
}
\description{
It retrieves all chromosome names having longer than 2.5Mbp.
}
\details{
To extract all chromosome names from .hic file
}
\examples{
myhic=system.file('extdata', 'example.hic', package ='HiCocietyExample')
get_all_chr_names(myhic)
}
\author{
Sora Yoon, PhD
}
