% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_all_chr_names}
\alias{get_all_chr_names}
\title{Retrieve chromosome names from .hic file}
\usage{
get_all_chr_names(fname)
}
\arguments{
\item{fname}{Path to .hic file}
}
\description{
It retrieves all chromosome names having longer than 2.5Mbp.
}
\details{
To extract all chromosome names from .hic file
}
\examples{
# Not run
# myhic=get_example_hic_file();
# get_all_chr_names(myhic)
}
\author{
Sora Yoon, PhD
}
