% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_txdb}
\alias{get_txdb}
\title{All available Txdb}
\usage{
get_txdb()
}
\value{
A character vector containing the names of all available TxDb packages.
}
\description{
It finds all available Txdb packages used in add_Genes function.
}
\details{
Check all available Txdb package
}
\examples{
get_txdb()
}
\author{
Sora Yoon, PhD
}
