% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getContactFrequency}
\alias{getContactFrequency}
\title{Get contact Frequency}
\usage{
getContactFrequency(fname, chr, resol)
}
\arguments{
\item{fname}{.hic data for any types of genome conformation capture data}

\item{chr}{Chromosome number of network extraction}

\item{resol}{DNA basepair Resolution. Default=10000}
}
\value{
A \code{data.frame} containing three columns: \code{x} and \code{y} (genomic coordinate pairs), and \code{counts} (the contact frequency between them).
}
\description{
Retrieve contact frequency from .hic file using strawR package.
}
\details{
Get Contact Frequncy from .hic file
}
\examples{
myhic=system.file('extdata', 'example.hic', package ='HiCocietyExample')
A = getContactFrequency(myhic,19,5000)
head(print(A))
}
\author{
Sora Yoon, PhD
}
