% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getContactFrequency}
\alias{getContactFrequency}
\title{Get contact Frequency}
\usage{
getContactFrequency(fname, chr, resol)
}
\arguments{
\item{fname}{.hic data for any types of genome conformation capture data}

\item{chr}{Chromosome number of network extraction}

\item{resol}{DNA basepair Resolution. Default=10000}
}
\description{
Retrieve contact frequency from .hic file using strawR package.
}
\details{
Get Contact Frequncy from .hic file
}
\examples{
# Not run
# myhic=get_example_hic_file()
# getContactFrequency(myhic,19,5000)
}
\author{
Sora Yoon, PhD
}
