% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getContactProbability}
\alias{getContactProbability}
\title{Contact probability}
\usage{
getContactProbability(tab, farthest = 2000000, resol = 10000, prob)
}
\arguments{
\item{tab}{Output from getContactFrequency function.}

\item{farthest}{Maximum 1-D distance to search. Default=2Mb}

\item{resol}{Hi-C resolution for test. Default = 10000}

\item{prob}{Significance cutoff for negative binomial distribution. Default =0.975}
}
\description{
It estimates contact probability based on the distance of a pair of a loci.
}
\details{
Get Contact probablity
}
\examples{
# Not run
# myhic=get_example_hic_file();
# mydf=getContactFrequency(myhic, 19, 5000);
# myprob=getContactProbability(mydf,farthest=2000000, resol=5000,prob=0.975);
}
\author{
Sora Yoon, PhD
}
