% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hic2community}
\alias{hic2community}
\title{Create module objects from the Hi-C data}
\usage{
hic2community(
  fname,
  chr,
  resol,
  nbprob,
  farthest,
  par.noise = 1,
  network.cluster.method = "louvain"
)
}
\arguments{
\item{fname}{Path to .hic file}

\item{chr}{chromosome numbersto run.}

\item{resol}{Resolution of Hi-C data}

\item{nbprob}{Negative binomial probability. Higher value gives smaller number of stronger interaction.}

\item{farthest}{The maximum searching distance between two nodes}

\item{par.noise}{Parameter for noise removal. Default is 1, higher value gives more filtered interactions.}

\item{network.cluster.method}{Can select between 'louvain' as default and 'label.prop' which means the label propagation method.}
}
\description{
It generates a list of graph of significant interactions, module table and module elements.
}
\details{
It generates a list of graph of significant interactions, module table and module elements.
}
\examples{
# Not run
# myhic = get_example_hic_file();
# mycom = hic2community(myhic, "19", 5000, 0.975, 2000000, par.noise=1, 'louvain')
}
\author{
Sora Yoon, PhD
}
