% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hic2network}
\alias{hic2network}
\title{HiC to network data format}
\usage{
hic2network(ftab)
}
\arguments{
\item{ftab}{three-column data composed of locus1, locus2 and value}
}
\description{
It converts Hi-C dataframe to network object.
}
\details{
Convert HiC to network data format
}
\examples{
# Not run
# myhic=get_example_hic_file();
# ftab=getContactFrequency(paste0('./',myhic),19,5000);
# net = hic2network(ftab[100:200,]);
# plot(net)
}
\author{
Sora Yoon, PhD
}
